Giza Device Whitepaper V. 1.2.1
There are risks, and uncertainties associated with GIZA and/or the
Distributor and their respective businesses and operations, the GIZA
Tokens, the GIZA Token Distribution Event, the Device App, and the
Company’s software, (each as referred to in this Whitepaper).
This Whitepaper, any part thereof and any copy thereof must not be
taken or transmitted to any country where distribution or dissemination
of Token Sale or Initial Coin Offering like the one described in this
Whitepaper is prohibited or restricted.
The GIZA Tokens are not intended to constitute securities in any
jurisdiction. GIZA Tokens are utility token and cannot have a performance
or a particular value.
Therefore, Therefore, this Whitepaper cannot constitute a prospectus or offer
document for investment in securities.
This Whitepaper does not constitute or form part of any opinion on any
advice to sell, or any solicitation of any offer by GIZA Device Ltd. to
purchase any GIZA Tokens or give any help in any investment
decision.
You are not eligible, and you are not offered to purchase any GIZA
Tokens in the GIZA Token Distribution Event (as referred to in this
Whitepaper) if you are a citizen, resident (tax or otherwise) or green card
holder of the United States of America or People's Republic of China.
This Whitepaper is for discussion and pre-information purposes only,
provided as a courtesy. The information contained herein is subject to
change, and GIZA Device Ltd is not obligated to inform anyone about the
change, adjustment or modification of this document. No part of this
document is legally binding or enforceable, nor is meant to be. You are
not eligible to not copy or disseminate any part of this document without
including this disclaimer.